requirements.txt
faust-streaming>=0.8.9
python-schema-registry-client>=2.4.1
pydantic[dotenv]>=0.21.0
loguru>=0.6.0
pytz>=2022.2.1
requests>=2.28.1
kaitaistruct>=0.10
simplejson>=3.17.6
taospy>=2.6.4